+++ src/surrealist/ql/statements/utils.py
@@ -1,5 +1,6 @@
-import json
 from typing import Optional, Dict
+
+from surrealist.utils import safe_dumps
 
 
 def combine(result: Optional[str], kwargs: Dict) -> str:
@@ -12,6 +13,6 @@ def combine(result: Optional[str], kwargs: Dict) -> str:
     :return: string combined
     """
     first = result if result else ''
-    second = ", ".join(f"{k} = {json.dumps(v)}" for k, v in kwargs.items()) if kwargs else ''
+    second = ", ".join(f"{k} = {safe_dumps(v)}" for k, v in kwargs.items()) if kwargs else ''
     args = ', '.join(element for element in (first, second) if element)
     return args